requirements.txt
python-dotenv==1.0.0
pyperclip==1.8.2
openai==1.63.2
